requirements.txt
Django>=5.0  
djangorestframework>=3.0  
djangorestframework-simplejwt==5.3.1  
pyjwt==2.10.1  
